[egs] add `tuning` folder
